# [HUDI-69] Support Spark Datasource for MOR table - RDD approach