--- a/hudi-spark/src/main/scala/org/apache/hudi/DefaultSource.scala
+++ b/hudi-spark/src/main/scala/org/apache/hudi/DefaultSource.scala
@@ -132,15 +132,11 @@ class DefaultSource extends RelationProvider
 
     log.info("Constructing hoodie (as parquet) data source with options :" + optParams)
     // simply return as a regular parquet relation
-    val relation =  DataSource.apply(
+    DataSource.apply(
       sparkSession = sqlContext.sparkSession,
       userSpecifiedSchema = Option(schema),
       className = "parquet",
       options = optParams)
       .resolveRelation()
-
-    sqlContext.sparkContext.hadoopConfiguration.unset("mapreduce.input.pathFilter.class")
-
-    relation
   }
 }