@umehrot2 2020-07-22
> @umehrot2 can you also please make a quick second pass.

@vinothchandar @garyli1019 Sorry for the late response. Plan to review it later today.