@garyli1019 2020-07-26
Added support for `PruneFilterScan`. Please review this PR again. Thank you!