@garyli1019 2020-07-29
@bvaradar Thanks for trying this out. `java.lang.NoSuchMethodError: org.apache.spark.sql.execution.datasources.PartitionedFile` looks strange. I will try it out on my production today to see if I can reproduce.