@garyli1019 2020-07-30
@bvaradar I tested on Spark 2.4.0 cdh release with a small dataset, and found a broadcast configuration issue. Pushed a new commit with the fix. Now this work fine on my cluster. I will test a larger dataset tomorrow.
I couldn't reproduce `java.lang.NoSuchMethodError: org.apache.spark.sql.execution.datasources.PartitionedFile`. Is this class somehow missing on aws release? Are you able to import from the spark-shell?