@vinothchandar 2020-08-04
@garyli1019 thanks for testing this! Will review again.

can we resolve the addressed comments in this PR, and also do a small summary on what the follow up work here is? 
custom payload support? (sorry if I am re-asking this, just switching back to this again)