@garyli1019 2020-08-04
> small summary on what the follow-up work here is?

@vinothchandar For 0.6.0 release, the only one left is incremental pulling. I am currently working on it and will probably get it done by tomorrow. Maybe we can wait until tomorrow to review the whole thing.
The vectorized reader and pruning are supported in the current version.

> custom payload support?

Yes, will let the `HoodieMergedLogRecordScanner` handle the payload loading and only need to specify `merge` or `skipmerge` when running the query. Included unit test for delete of `OverwriteWithLatestAvroPayload` to verify custom payload mechanism.