@vinothchandar 2020-08-04
>Yes, will let the HoodieMergedLogRecordScanner handle the payload loading and only need to specify merge or skipmerge when running the query. Included unit test for delete of OverwriteWithLatestAvroPayload to verify custom payload mechanism.


to be clear, you are saying it should all be working correct? assuming you may not have conflicts with #1807 , can you please rebase this off latest masteR?