@garyli1019 2020-08-04
> to be clear, you are saying it should all be working correct? assuming you may not have conflicts with #1807 , can you please rebase this off latest masteR?

Yes the custom payload is working. Done rebasing.